backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlmodel>=0.0.16
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
slowapi>=0.1.9
cachetools>=5.3.0
openai-agents>=0.6.0
cohere>=5.5.3